Implement quota for all users
Because too many users where making too many requests that made it
un-sustainable to keep the API free for everybody else. Therefore I
implemented hard quota limits so that this API will stay free.

(some people have paid to have bigger quotas because of a lack of
programming prowess to implement local caching of the request on their
side)